Update usage of feature gate
